+++ app/lib/screens/storage/utils/delete_run.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+import 'package:profit_taker_analyzer/screens/storage/model/run_list_model.dart';
+import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:rust_core/rust_core.dart';
+
+Future<bool> deleteRun(BuildContext context, RunListItemCustom run) async {
+  final scaffoldMessenger = ScaffoldMessenger.of(context);
+
+  // Show a confirmation dialog before deleting
+  final bool? confirmDelete = await showDialog<bool>(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text(FlutterI18n.translate(context, "alerts.delete_title")),
+        content: Text(FlutterI18n.translate(context, "alerts.delete_message")),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(context).pop(false),
+            child: Text(FlutterI18n.translate(context, "buttons.cancel")),
+          ),
+          TextButton(
+            onPressed: () => Navigator.of(context).pop(true),
+            child: Text(FlutterI18n.translate(context, "buttons.delete")),
+          ),
+        ],
+      );
+    },
+  );
+
+  if (confirmDelete == true) {
+    // Call the Rust function to delete the run
+    final DeleteRunResult result = deleteRunFromDb(runId: run.id);
+
+    if (result.success) {
+      // Show success message
+      if (context.mounted) {
+        scaffoldMessenger.showSnackBar(
+          SnackBar(
+              content:
+                  Text(FlutterI18n.translate(context, "delete_run.success"))),
+        );
+      }
+      return true; // Return true to indicate success
+    } else {
+      // Show error message
+      if (context.mounted) {
+        scaffoldMessenger.showSnackBar(
+          SnackBar(
+            content: Text(result.error ??
+                FlutterI18n.translate(context, "delete_run.error")),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+      return false; // Return false to indicate failure
+    }
+  }
+
+  return false; // Return false if deletion was canceled
+}
